requirements.txt
pydantic>=2.0.0
jsonlines
python-dotenv
openai
trafilatura
httpx
requests
